# services/auth/pkg/swagger/userApi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "contact": {
      "name": "Evgenia Kivotova",
      "email": "e.kivotova@gmail.com"
    }
  },
  "tags": [
    {
      "name": "UserV1"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/v1": {
      "get": {
        "summary": "Get user",
        "operationId": "UserV1_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1GetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "delete": {
        "summary": "Delete user",
        "operationId": "UserV1_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "patch": {
        "summary": "Update user data",
        "operationId": "UserV1_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_v1UpdateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    },
    "/user/v1/create": {
      "post": {
        "summary": "Create new user",
        "operationId": "UserV1_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_v1CreateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    },
    "/user/v1/list": {
      "get": {
        "summary": "List users with filters",
        "operationId": "UserV1_GetList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1GetListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filters.names",
            "description": "List of unique user names to filter",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "user_v1CreateRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "title": "Password of a user"
        },
        "passwordConfirm": {
          "type": "string",
          "title": "Repeated password to prevent user typos"
        },
        "info": {
          "$ref": "#/definitions/user_v1UserInfo"
        }
      }
    },
    "user_v1CreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "user_v1GetListFilters": {
      "type": "object",
      "properties": {
        "names": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "List of unique user names to filter"
        }
      }
    },
    "user_v1GetListResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/user_v1User"
          }
        }
      }
    },
    "user_v1GetResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/user_v1User"
        }
      }
    },
    "user_v1UpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/user_v1UserRole"
        }
      }
    },
    "user_v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Internal user id"
        },
        "info": {
          "$ref": "#/definitions/user_v1UserInfo"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "title": "Time when user was created"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "title": "Time when user was last updated"
        }
      }
    },
    "user_v1UserInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Unique name of a user"
        },
        "email": {
          "type": "string",
          "title": "Email of a user"
        },
        "role": {
          "$ref": "#/definitions/user_v1UserRole"
        }
      }
    },
    "user_v1UserRole": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "USER",
        "ADMIN"
      ],
      "default": "UNKNOWN"
    }
  }
}